+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added the `notifications.notify-self` option. When set to `true`, the bot will also send a
   a notification to the user that just joined (if he is registered). By default, this is set to false.
 - The bot will now remove a server and all its registrations from the database when it is removed from the server.
+- Add the `/voice-watch status` command to check the registration status of a user.
 
 ### Changed
 - Styled the messages that are sent to users